part_3/app/requirements.txt
azure-functions==1.21.1
